up to 2013-02-03

+++ dokuwiki-plugin-nssize.spec
@@ -1,17 +1,21 @@
+%define		subver	2013-02-03
+%define		ver		%(echo %{subver} | tr -d -)
 %define		plugin		nssize
+%define		php_min_version 5.3.0
+%include	/usr/lib/rpm/macros.php
 Summary:	DokuWiki plugin to Show disk usage of a namespace
 Name:		dokuwiki-plugin-%{plugin}
-Version:	20080322
-Release:	0.2
+Version:	%{ver}
+Release:	1
 License:	GPL v2
 Group:		Applications/WWW
-Source0:	http://wiki.symplus.com/_media/computer/source/nssize.zip
-# Source0-md5:	288b06bc6765f9b51e5a3aa11b99b3b7
-URL:		http://www.dokuwiki.org/plugin:nssize
-Patch0:		doku-conf.patch
+Source0:	https://github.com/iobataya/dokuwiki-plugin-nssize/archive/6bccd121/%{plugin}-%{subver}.tar.gz
+# Source0-md5:	8b3d5f57c2e2045e92a593464bd08022
+URL:		https://www.dokuwiki.org/plugin:nssize
 BuildRequires:	rpmbuild(macros) >= 1.553
 BuildRequires:	unzip
 Requires:	dokuwiki >= 20061106
+Requires:	php(core) >= %{php_min_version}
 BuildArch:	noarch
 BuildRoot:	%{tmpdir}/%{name}-%{version}-root-%(id -u -n)
 
@@ -30,14 +34,20 @@ selectable in the plugin-configuration page.
 
 %prep
 %setup -qc
-mv %{plugin}/* .
-%undos *.css *.php
-%patch0 -p1
+mv *-%{plugin}-*/* .
+
+%build
+version=$(awk '/^date/{print $2}' plugin.info.txt)
+if [ "$(echo "$version" | tr -d -)" != %{version} ]; then
+	: %%{version} mismatch
+	exit 1
+fi
 
 %install
 rm -rf $RPM_BUILD_ROOT
 install -d $RPM_BUILD_ROOT%{plugindir}
 cp -a . $RPM_BUILD_ROOT%{plugindir}
+%{__rm} $RPM_BUILD_ROOT%{plugindir}/README
 
 # find locales
 %find_lang %{name}.lang
@@ -45,15 +55,10 @@ cp -a . $RPM_BUILD_ROOT%{plugindir}
 %clean
 rm -rf $RPM_BUILD_ROOT
 
-%post
-# force css cache refresh
-if [ -f %{dokuconf}/local.php ]; then
-	touch %{dokuconf}/local.php
-fi
-
 %files -f %{name}.lang
 %defattr(644,root,root,755)
+%doc README
 %dir %{plugindir}
 %{plugindir}/*.php
-%{plugindir}/*.css
+%{plugindir}/*.txt
 %{plugindir}/conf
